fix: move trajectory_span_model
