+++ src/uipath/eval/evaluators/trajectory_evaluator.py
@@ -6,12 +6,16 @@ from typing import Any, Optional
 from opentelemetry.sdk.trace import ReadableSpan
 from pydantic import field_validator
 
-from uipath._cli._evals._models._trajectory_span import TrajectoryEvaluationTrace
 from uipath.eval.models import EvaluationResult
 
 from ..._services import UiPathLlmChatService
 from ..._utils.constants import COMMUNITY_agents_SUFFIX
-from ..models.models import AgentExecution, LLMResponse, NumericEvaluationResult
+from ..models.models import (
+    AgentExecution,
+    LLMResponse,
+    NumericEvaluationResult,
+    TrajectoryEvaluationTrace,
+)
 from .base_evaluator import BaseEvaluator
